refactor(nika-cli): split the check render surface out of the verb module
ADR-023: NEP-0002's TRIFECTA rung pushed check/mod.rs to 1,507 LOC over
the 1,500 file ceiling (ratchet loc-limits). The themed human render
sections + row builders move to check/render.rs (746 LOC); the machine
(--json) surface, run/run_many/run_infer_permits and the tests stay.
Pure move — no behaviour change, E2E fixtures re-verified.
